+++ R/icd_to_ccsr.R
@@ -118,7 +118,7 @@
 #' drv <- dbDriver("PostgreSQL")
 #' dbcon <- DBI::dbConnect(drv,
 #'                         dbname = "db",
-#'                         host = "172.XX.XX.XXX",
+#'                         host = "domain_name.ca",
 #'                         port = 1234,
 #'                         user = getPass("Enter user:"),
 #'                         password = getPass("password"))
@@ -128,10 +128,7 @@
 #' }
 icd_to_ccsr <- function(dbcon, dxtable, type_mrdx = TRUE, unique_mrdx = FALSE, replace_invalidpdx = TRUE) {
 
-  cat("\n***Note:***
-  The output of this function is based on GEMINI-derived mappings of ICD-10-CA codes to CCSR categories.
-  Please carefully check mapping coverage for your cohort of interest, or contact the GEMINI team if you require additional support.\n")
-
+  mapping_message("ICD-10-CA codes to CCSR categories")
 
   cat(paste0(
     "\nObtaining CCSR categories for ICD-10-CA codes in input table ",
@@ -139,9 +136,7 @@ icd_to_ccsr <- function(dbcon, dxtable, type_mrdx = TRUE, unique_mrdx = FALSE, r
 
   #######  Check user inputs  #######
   ## Valid DB connection?
-  if (!isPostgresqlIdCurrent(dbcon)) {
-    stop("Invalid user input for argument dbcon. Please input a valid database connection.")
-  }
+  check_input(dbcon, "DBI")
 
   ## dxtable provided as data.frame/data.table?
   if (!any(class(dxtable) %in% c("data.frame", "data.table"))) {